+++ events/messageCreate.js
@@ -1,43 +1,42 @@
-const { MessageEmbed, Collection } = require('..');
-const Keyv = require('keyv');
-const prefixes = new Keyv(`${process.env.MONGO_URL}`);
+const { Message, MessageEmbed, Collection } = require('..');
+const config = require("../configs/config.json");
+var ee = require("../configs/embed.json");
 const client = require("..");
-//var ee = require("../configs/embed.json");
-const config = require('../configs/config.json');
-const globalPrefix = config.globalPrefix;
+const prefix = config.globalPrefix
 
-client.on('messageCreate', async (message) => {
-	//const { escapeRegex, onCoolDown } = require("../utils/function.js");
-// Ignore WebManagebot, DMs, partial messages in channels and themselves
-	if (message.author.bot) return;
-	if (!message.guild) return message.reply('Please send commands in a server');
-	if (message.channel.partial) await message.channel.fetch();
-	if (message.partial) await message.fetch();
+client.on("messageCreate", async (message) => {
+  const { escapeRegex, onCoolDown } = require("../utils/function.js");
+  if (!message.guild) return;
+  if (message.author.bot) return;
+  if (message.channel.partial) await message.channel.fetch();
+  if (message.partial) await message.fetch();
+  const prefixRegex = new RegExp(
+    `^(<@!?${client.user.id}>|${escapeRegex(prefix)})\\s*`
+  );
+  if (!prefixRegex.test(message.content)) return;
+  const [, matchedPrefix] = message.content.match(prefixRegex);
+  const args = message.content.slice(matchedPrefix.length).trim().split(/ +/);
+  const cmd = args.shift().toLowerCase();
+// Retrieve mention prefix
+  if (cmd.length === 0) {
+    if (matchedPrefix.includes(client.user.id)) {
+      message.reply(
+        `To see all commands type: ${prefix}help`
+      );
+    }
+  }
+  const command = client.commands.get(cmd.toLowerCase());
+  if (!command) return;
+  if (command) {
+    if (!message.member.permissions.has(command.permissions || []))
+      return message.channel.send('You do not have the perms to complete such actions...');
 
-	let args;
-// Handle messages in a guild
-	if (message.guild) {
-		let prefix;
-
-		if (message.content.startsWith(globalPrefix)) {
-			prefix = globalPrefix;
-		} else {
-			// check the guild-level prefix
-			const guildPrefix = await prefixes.get(message.guild.id);
-			if (message.content.startsWith(guildPrefix)) prefix = guildPrefix;
-		}
-
-// If we found a prefix, setup args; otherwise, this isn't a command
-		if (!prefix) return;
-		args = message.content.slice(prefix.length).trim().split(/\s+/);
-	} else {
-// Handle DMs
-		const slice = message.content.startsWith(globalPrefix) ? globalPrefix.length : 0;
-		args = message.content.slice(slice).split(/\s+/);
-	}
-
-// Get the first space-delimited argument after the prefix as the command
-	const command = args.shift().toLowerCase();
-// Send the command
-  await command.run(client, message, args);
+  // Check if user is on cooldown with the cmd, with Tomato#6966's Function from /handlers/functions.js
+    if (onCoolDown(message, command)) {
+      let cool = new MessageEmbed()
+      .setDescription(`❌ Please wait ${onCoolDown(message, command)} more second(s) before reusing the ${command.name} command.`)
+      return message.channel.send({embeds : [cool]})
+    }
+    await command.run(client, message, args, prefix);
+  };
 });
